컴포넌트로 나눔
